requirements.txt
streamlit
python-dotenv
pinecone-client
langchain-openai
langchain-google-genai
langchain-mistralai
boto3
urllib3
